moved operators.go into primitives, where implementation details are apparent
